+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - Support for steps to navigate to different slugs depending on the state of the
   step
 - Support for specifying dynamic keys for `ComponentDatabaseMap`s
+- Support for marking `DynamicComponent`s as required
 
 ### Changed